# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserLevel = typeof UserLevel[keyof typeof UserLevel];


export const UserLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export type UserSkillCategory = typeof UserSkillCategory[keyof typeof UserSkillCategory];


export const UserSkillCategory = {
  backend: 'backend',
  frontend: 'frontend',
  mobile: 'mobile',
  ai: 'ai',
  networking: 'networking',
  design: 'design',
  devops: 'devops',
  other: 'other',
} as const;

export type UserSkillProficiency = typeof UserSkillProficiency[keyof typeof UserSkillProficiency];


export const UserSkillProficiency = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export interface UserSkill {
  name: string;
  category: UserSkillCategory;
  proficiency: UserSkillProficiency;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  skills: UserSkill[];
  level: UserLevel;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  lookingFor?: string | null;
  roomsJoined: number;
  postsCount: number;
  createdAt: string;
}

export type UserInputLevel = typeof UserInputLevel[keyof typeof UserInputLevel];


export const UserInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export interface UserInput {
  username: string;
  displayName: string;
  bio?: string;
  avatarUrl?: string;
  location?: string;
  skills: UserSkill[];
  level: UserInputLevel;
  githubUrl?: string;
  twitterUrl?: string;
  lookingFor?: string;
}

export type UserUpdateLevel = typeof UserUpdateLevel[keyof typeof UserUpdateLevel];


export const UserUpdateLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  location?: string;
  skills?: UserSkill[];
  level?: UserUpdateLevel;
  githubUrl?: string;
  twitterUrl?: string;
  lookingFor?: string;
}

export type UserStatsByLevel = {
  beginner: number;
  intermediate: number;
  pro: number;
};

export type UserStatsByCategoryItem = {
  category: string;
  count: number;
};

export interface UserStats {
  totalUsers: number;
  totalRooms: number;
  totalPosts: number;
  byLevel: UserStatsByLevel;
  byCategory: UserStatsByCategoryItem[];
}

export type RoomType = typeof RoomType[keyof typeof RoomType];


export const RoomType = {
  discussion: 'discussion',
  project: 'project',
  learning: 'learning',
} as const;

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  type: RoomType;
  skills: string[];
  memberCount: number;
  postCount: number;
  isPinned: boolean;
  isPrivate: boolean;
  /** @nullable */
  createdByUserId?: number | null;
  createdAt: string;
  lastActiveAt: string;
}

export type RoomInputType = typeof RoomInputType[keyof typeof RoomInputType];


export const RoomInputType = {
  discussion: 'discussion',
  project: 'project',
  learning: 'learning',
} as const;

export interface RoomInput {
  name: string;
  description?: string;
  type: RoomInputType;
  skills: string[];
  isPrivate?: boolean;
  createdByUserId?: number;
}

export interface RoomUpdate {
  name?: string;
  description?: string;
  skills?: string[];
  isPinned?: boolean;
  isPrivate?: boolean;
}

export interface Post {
  id: number;
  roomId: number;
  /** @nullable */
  authorId?: number | null;
  author?: User;
  content: string;
  isPinned: boolean;
  upvotes: number;
  /** @nullable */
  parentPostId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface RoomActivity {
  room: Room;
  recentPostCount: number;
  lastPost?: Post;
}

export interface PostInput {
  content: string;
  authorId?: number;
  parentPostId?: number;
}

export interface PostUpdate {
  content?: string;
  isPinned?: boolean;
}

export type MatchCriteriaLevel = typeof MatchCriteriaLevel[keyof typeof MatchCriteriaLevel];


export const MatchCriteriaLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export interface MatchCriteria {
  skills: string[];
  level?: MatchCriteriaLevel;
  location?: string;
  lookingFor?: string;
  limit?: number;
}

export interface MatchResult {
  user: User;
  matchScore: number;
  matchReasons: string[];
}

export interface AiSummary {
  roomId: number;
  summary: string;
  keyTopics: string[];
  generatedAt: string;
}

export interface SolutionRequest {
  problem: string;
  context?: string;
  skills?: string[];
}

export interface AiSolution {
  solution: string;
  steps: string[];
  resources: string[];
}

export type ListUsersParams = {
skill?: string;
level?: ListUsersLevel;
location?: string;
limit?: number;
offset?: number;
};

export type ListUsersLevel = typeof ListUsersLevel[keyof typeof ListUsersLevel];


export const ListUsersLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  pro: 'pro',
} as const;

export type GetTopBuildersParams = {
limit?: number;
};

export type ListRoomsParams = {
type?: ListRoomsType;
skill?: string;
limit?: number;
offset?: number;
};

export type ListRoomsType = typeof ListRoomsType[keyof typeof ListRoomsType];


export const ListRoomsType = {
  discussion: 'discussion',
  project: 'project',
  learning: 'learning',
} as const;

export type GetLiveRoomActivityParams = {
limit?: number;
};

export type GetRoomPostsParams = {
roomId: number;
limit?: number;
offset?: number;
};