# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DevLink KE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [discussion, project, learning]
        - name: skill
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room (increment member count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /posts:
    get:
      operationId: getRoomPosts
      tags: [posts]
      summary: List posts in a room
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /rooms/{id}/posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post in a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/upvote:
    post:
      operationId: upvotePost
      tags: [posts]
      summary: Upvote a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upvoted post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /match/developers:
    post:
      operationId: matchDevelopers
      tags: [matching]
      summary: Find developers matching a set of criteria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchCriteria"
      responses:
        "200":
          description: Matched developers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
  /match/suggestions/{userId}:
    get:
      operationId: getMatchSuggestions
      tags: [matching]
      summary: AI-powered match suggestions for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
  /ai/summarize/{roomId}:
    get:
      operationId: summarizeRoom
      tags: [ai]
      summary: AI summary of recent room discussion
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"
  /ai/suggest-solution:
    post:
      operationId: suggestSolution
      tags: [ai]
      summary: AI-powered solution suggestion for a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionRequest"
      responses:
        "200":
          description: AI solution suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSolution"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/UserSkill"
        level:
          type: string
          enum: [beginner, intermediate, pro]
        githubUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        lookingFor:
          type: ["string", "null"]
        roomsJoined:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, username, displayName, skills, level, roomsJoined, postsCount, createdAt]
    UserSkill:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [backend, frontend, mobile, ai, networking, design, devops, other]
        proficiency:
          type: string
          enum: [beginner, intermediate, pro]
      required: [name, category, proficiency]
...
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [discussion, project, learning]
        skills:
          type: array
          items:
            type: string
        memberCount:
          type: integer
        postCount:
          type: integer
        isPinned:
          type: boolean
        isPrivate:
          type: boolean
        createdByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
      required: [id, name, type, skills, memberCount, postCount, isPinned, isPrivate, createdAt, lastActiveAt]
    RoomInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [discussion, project, learning]
        skills:
          type: array
          items:
            type: string
        isPrivate:
          type: boolean
        createdByUserId:
          type: integer
      required: [name, type, skills]
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        isPrivate:
          type: boolean
...
    Post:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        authorId:
          type: ["integer", "null"]
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        isPinned:
          type: boolean
        upvotes:
          type: integer
        parentPostId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, roomId, content, isPinned, upvotes, createdAt, updatedAt]
    PostInput:
      type: object
      properties:
        content:
          type: string
        authorId:
          type: integer
        parentPostId:
          type: integer
      required: [content]
    PostUpdate:
      type: object
      properties:
        content:
          type: string
        isPinned:
          type: boolean
    MatchCriteria:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        level:
          type: string
          enum: [beginner, intermediate, pro]
        location:
          type: string
        lookingFor:
          type: string
        limit:
          type: integer
          default: 10
      required: [skills]
    MatchResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        matchScore:
          type: number
        matchReasons:
          type: array
          items:
            type: string
      required: [user, matchScore, matchReasons]
    AiSummary:
      type: object
      properties:
        roomId:
          type: integer
        summary:
          type: string
        keyTopics:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
      required: [roomId, summary, keyTopics, generatedAt]
    SolutionRequest:
      type: object
      properties:
        problem:
          type: string
        context:
          type: string
        skills:
          type: array
          items:
            type: string
      required: [problem]
    AiSolution:
      type: object
      properties:
        solution:
          type: string
        steps:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            type: string
      required: [solution, steps, resources]