# lib/api-zod/src/generated/types/aiSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */

export interface AiSummary {
  roomId: number;
  summary: string;
  keyTopics: string[];
  generatedAt: Date;
}