# lib/api-zod/src/generated/types/matchCriteria.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchCriteriaLevel } from './matchCriteriaLevel';

export interface MatchCriteria {
  skills: string[];
  level?: MatchCriteriaLevel;
  location?: string;
  lookingFor?: string;
  limit?: number;
}