# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomType } from './roomType';

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  type: RoomType;
  skills: string[];
  memberCount: number;
  postCount: number;
  isPinned: boolean;
  isPrivate: boolean;
  /** @nullable */
  createdByUserId?: number | null;
  createdAt: Date;
  lastActiveAt: Date;
}