# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */
import type { UserLevel } from './userLevel';
import type { UserSkill } from './userSkill';

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  skills: UserSkill[];
  level: UserLevel;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  lookingFor?: string | null;
  roomsJoined: number;
  postsCount: number;
  createdAt: Date;
}