# lib/db/src/schema/posts.ts
import { pgTable, text, serial, integer, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const postsTable = pgTable("posts", {
  id: serial("id").primaryKey(),
  roomId: integer("room_id").notNull(),
  authorId: integer("author_id"),
  content: text("content").notNull(),
  isPinned: boolean("is_pinned").notNull().default(false),
  upvotes: integer("upvotes").notNull().default(0),
  parentPostId: integer("parent_post_id"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertPostSchema = createInsertSchema(postsTable).omit({ id: true, isPinned: true, upvotes: true, createdAt: true, updatedAt: true });
export type InsertPost = z.infer<typeof insertPostSchema>;
export type Post = typeof postsTable.$inferSelect;